requirements.txt
# GrantBot dependencies

# HTTP requests
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
openai>=1.0.0
anthropic>=0.25.0
python-dateutil>=2.8.2
pytz>=2023.3
markdownify>=0.11.6
markdown2>=2.4.0